+++ src/helpers/app-error.helper.ts
@@ -1,0 +1,13 @@
+export class AppError {
+  public readonly message: string;
+
+  public readonly statusCode: number;
+
+  public readonly details?: any;
+
+  constructor(message: any, statusCode = 400, details?: any) {
+    this.message = message;
+    this.statusCode = statusCode;
+    this.details = details;
+  }
+}