+++ src/index.ts
@@ -1,10 +1,3 @@
-import express, { Application, Request, Response, NextFunction } from 'express';
+import { StartupConfiguration } from './configs/startup.config';
 
-const app: Application = express();
-const port = 5000;
-
-app.use('/', (_req: Request, res: Response, _next: NextFunction) => {
-  res.status(200).send({ data: 'Hello from Ornio AS' });
-});
-
-app.listen(port, () => console.log(`Server is listening on port ${port}!`));
+StartupConfiguration.run();